Use ruamel_yaml instead of PyYAML
PyYAML was loading float values expressed in scientific notation as
strings (see, e.g., https://stackoverflow.com/q/30458977/1563298),
with caused BRaKE to fail. ruamel_yaml loads the values correctly.
It's available on PyPI and conda, and installed by default with
Miniconda, so it seems fairly well-accepted. Note that this issue may
be OS-related, as PyYAML worked correctly on macOS but not CentOS.